requirements.txt
chainlit==0.7.700
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.1
pydantic>=2.0 